revert layer_norm opt

+++ paddle/phi/infermeta/spmd_rules/layer_norm.h
@@ -23,7 +23,7 @@ namespace distributed {
 SpmdInfo LayerNormInferSpmd(const DistMetaTensor& x,
                             const DistMetaTensor& scale,
                             const DistMetaTensor& bias,
-                            double epsilon,
+                            float epsilon,
                             int begin_norm_axis);
 
 SpmdInfo LayerNormGradInferSpmd(const DistMetaTensor& x,
@@ -32,7 +32,7 @@ SpmdInfo LayerNormGradInferSpmd(const DistMetaTensor& x,
                                 const DistMetaTensor& mean,
                                 const DistMetaTensor& variance,
                                 const DistMetaTensor out_grad,
-                                double epsilon = 1e-5,
+                                float epsilon = 1e-5,
                                 int begin_norm_axis = 1);
 
 SpmdInfo LayerNormInferSpmdReverse(const DistMetaTensor& x,
@@ -41,20 +41,20 @@ SpmdInfo LayerNormInferSpmdReverse(const DistMetaTensor& x,
                                    const DistMetaTensor& out,
                                    const DistMetaTensor& mean,
                                    const DistMetaTensor& variance,
-                                   double epsilon,
+                                   float epsilon,
                                    int begin_norm_axis);
 
 SpmdInfo FastLnInferSpmd(const DistMetaTensor& x,
                          const DistMetaTensor& scale,
                          const DistMetaTensor& bias,
-                         double epsilon);
+                         float epsilon);
 
 SpmdInfo FastLnGradInferSpmd(const DistMetaTensor& x,
                              const DistMetaTensor& scale,
                              const DistMetaTensor& mean,
                              const DistMetaTensor& invvar,
                              const DistMetaTensor& y_grad,
-                             double epsilon);
+                             float epsilon);
 
 }  // namespace distributed
 }  // namespace phi
