+++ paddle/phi/kernels/cpu/layer_norm_kernel.cc
@@ -33,7 +33,7 @@ void LayerNormKernel(const Context& dev_ctx,
                      const DenseTensor& x,
                      const optional<DenseTensor>& scale_opt,
                      const optional<DenseTensor>& bias_opt,
-                     double epsilon,
+                     float epsilon,
                      int begin_norm_axis,
                      DenseTensor* y,
                      DenseTensor* mean,
@@ -141,7 +141,7 @@ void LayerNormKernel(const Context& dev_ctx,
       scale ? scale->data<T>() : nullptr,
       bias ? bias->data<T>() : nullptr,
       static_cast<int>(left),
-      static_cast<double>(epsilon),
+      static_cast<float>(epsilon),
       right);
 #endif
 }
